requirements.txt
openai>=1.55.0
python-decouple>=3.8
tiktoken>=0.8.0
python-dotenv>=1.0.1
uvicorn>=0.32.1
fastapi>=0.116.2
llama-index-core>=0.14.0
llama-index-graph-stores-neo4j>=0.4.0
llama-index-vector-stores-neo4jvector>=0.4.0
llama-index-llms-anthropic>=0.10.0
llama-index-llms-openai>=0.3.0
llama-index-embeddings-voyageai>=0.5.0
llama-index-embeddings-openai>=0.3.0
anthropic>=0.72.0
neo4j>=5.28.0
aiofiles>=25.1.0
tree-sitter==0.25.2
tree_sitter_python==0.25.0
jinja2==3.1.6